Add files via upload
Please note there is a bug in HumanoidNetworkPlayer that will need to be fixed in the next video, this bug is related to the avtBody transform and the avLeft and avRight transforms - the body isn't moving along if you eg. stand up and take a step in the real world where you hands/controllers do. I guess thats going to be the topic for episode 8 :)